tennis_video_analyzer/requirements.txt
# Dependencies for tennis_video_analyzer
guru4elephant-jsonflow>=0.1.1
jsonflow[all]>=0.1.0
opencv-python>=4.5.5
numpy>=1.20.0
Pillow>=9.0.0
requests>=2.27.1
matplotlib>=3.5.0
tqdm>=4.62.0 